+++ MarketPortal/Market/Data/Models/Order.cs
@@ -20,9 +20,9 @@ namespace Market.Data.Models
         public bool IsDenied { get; set; }
 
         [Required]
-        public double Quantity { get; set; }
+        public decimal Quantity { get; set; }
         [Required]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         [Required]
         public string Address { get; set; }
